# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreateAdminUserRequest:
    properties:
      email:
        type: string
      storageQuotaBytes:
        type: integer
      tempPassword:
        type: string
      username:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.SettingsResponse:
    properties:
      registrationEnabled:
        type: boolean
    type: object
...
  handlers.StatsResponse:
    properties:
      activeUsers:
        type: integer
      totalCollections:
        type: integer
      totalFiles:
        type: integer
      totalStorageUsedBytes:
        type: integer
      totalUsers:
        type: integer
    type: object
...
  handlers.UpdateAdminSettingsRequest:
    properties:
      registrationEnabled:
        type: boolean
    type: object
  handlers.UpdateAdminUserRequest:
    properties:
      isActive:
        type: boolean
      isAdmin:
        type: boolean
      storageQuotaBytes:
        type: integer
    type: object
...
  handlers.UserRow:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      isAdmin:
        type: boolean
      storageQuotaBytes:
        type: integer
      storageUsedBytes:
        type: integer
      totpEnabled:
        type: boolean
      username:
        type: string
    type: object
host: localhost
info:
  contact: {}
  description: Self-hosted, end-to-end encrypted file storage with federation. All
    file content and metadata are encrypted client-side; the server stores only ciphertext.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Kutup API
  version: 1.0.0
paths:
  /admin/settings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SettingsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get server settings
      tags:
      - Admin
    put:
      consumes:
      - application/json
      parameters:
      - description: Settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateAdminSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SettingsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update server settings
      tags:
      - Admin
  /admin/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StatsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get aggregate server statistics
      tags:
      - Admin
  /admin/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.UserRow'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - Admin
    post:
      consumes:
      - application/json
      parameters:
      - description: User details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateAdminUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Email or username already taken
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a user account (admin)
      tags:
      - Admin
  /admin/users/{id}:
    delete:
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: ""
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user and all their data
      tags:
      - Admin
    put:
      consumes:
      - application/json
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update (all optional)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateAdminUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a user (quota, active, admin flag)
      tags:
      - Admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"